third_party/blink/web_tests/http/tests/preload/warning/sync-request-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE WARNING: A preload for 'http://127.0.0.1:8000/preload/warning/resources/hello.txt' is found, but is not used because the new request is synchronous.
This is a testharness.js-based test.
PASS Preload match fails because the request is synchronous (see the console log).
Harness: the test ran to completion.
